# docs/api/desktop-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Knotra client engine boundary",
    "version": "1.0.0",
    "description": "Implemented HTTP API for the Go engine and CLI. The protocol identifier remains knotra.desktop/1. See desktop-v1.md for normative idempotency, events, security and execution rules."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787/v1",
      "description": "Local engine"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {}
  ],
  "paths": {
...
    "/artifacts": {
      "get": {
        "operationId": "get_artifacts",
        "summary": "List immutable registered artifact descriptors",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactPage"
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "post_artifacts",
        "summary": "Register uploaded input bytes",
        "responses": {
          "201": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifact": {
                      "$ref": "#/components/schemas/Artifact"
                    }
                  },
                  "required": [
                    "artifact"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "mediaType": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string",
                    "contentEncoding": "base64"
                  }
                },
                "required": [
                  "name",
                  "mediaType",
                  "content"
                ]
              }
            }
          }
        }
      }
    },
    "/artifacts/{artifactId}": {
      "get": {
        "operationId": "get_artifacts_artifactId",
        "summary": "Read artifact identity, hash and origin",
        "responses": {
          "200": {
            "description": "Successful result or durable command receipt.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifact": {
                      "$ref": "#/components/schemas/Artifact"
                    }
                  },
                  "required": [
                    "artifact"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Definitive rejection or engine failure. Consult status and code before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "artifactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            }
          }
        },
        "required": [
          "code",
          "message",
          "diagnostics"
        ]
      },
      "Diagnostic": {
        "type": "object",
        "properties": {
          "severity": {
            "enum": [
              "error",
              "warning"
            ]
          },
          "code": {
            "type": "string"
          },
          "phase": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "file": {
            "type": "string"
          },
          "line": {
            "type": "integer",
            "minimum": 1
          },
          "column": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "severity",
          "code",
          "phase",
          "message",
          "path"
        ]
      },
...
      "Artifact": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "name": {
            "type": "string"
          },
          "mediaType": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "minimum": 0,
            "maximum": 67108864
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "origin": {
            "type": "object",
            "properties": {
              "runId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              "instanceId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              "attemptId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              "operationId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              }
            },
            "required": []
          }
        },
        "required": [
          "id",
          "name",
          "mediaType",
          "size",
          "sha256",
          "origin"
        ]
      },
...
      "ArtifactPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Artifact"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Required for remote HTTPS deployments; optional loopback according to engine policy."
      }
    }
  }
}